+++ bld/clib/win386/c/escape.c
@@ -32,7 +32,7 @@
 
 #include "cover.h"
 
-short PASCAL _Cover_Escape( HDC dc, int a, int b, LPSTR c, LPSTR d )
+short PASCAL _Cover_Escape( HDC dc, short a, short b, LPSTR c, LPSTR d )
 {
     short       rc;
