+++ bld/gui/win/c/guicursr.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Set or reset mouse pointer type.
 *
 ****************************************************************************/
 
@@ -57,6 +56,6 @@ void *GUISetMouseCursor( gui_mouse_cursor type )
 void GUIResetMouseCursor( void *old )
 {
     if( old != NULL ) {
-            _wpi_setcursor( (HCURSOR)old );
+            _wpi_setcursor( (HCURSOR)(UINT)old );
     }
 }
