+++ bld/sdk/rc/rc/h/iortns.h
@@ -24,18 +24,17 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Client callback prototypes for wres library.
 *
 ****************************************************************************/
 
 
 int  RcOpen( const char *, int, ... );
 int  RcClose( int );
-int  RcWrite( int, const void *, int );
-int  RcRead( int, void *, int );
-long RcSeek( int, long, int );
-long RcTell( int );
+int  RcWrite( int, const void *, size_t );
+int  RcRead( int, void *, size_t );
+off_t RcSeek( int, off_t, int );
+off_t RcTell( int );
 
 void CloseAllFiles( void );
 void RegisterOpenFile( int fhdl );
