+++ bld/sdk/wde/h/wdefutil.h
@@ -48,39 +48,35 @@
 /****************************************************************************/
 /* function prototypes                                                      */
 /****************************************************************************/
-extern UINT     WdeTranslateNCMouseMsg   ( UINT );
-extern void     WdeSnapPointToGrid       ( POINT *pt );
-extern OBJPTR   WdeGetNextObject         ( Bool, OBJPTR, OBJPTR );
-extern OBJPTR   WdeCloneObject           ( OBJPTR, POINT * );
-extern void     WdeFreeControlIDs        ( void );
-extern uint_16  WdeGetNextControlID      ( void );
-extern uint_32  WdeHammingDistance       ( uint_32, uint_32 );
-extern Bool     WdeIsClassDefined        ( char * );
-extern char    *WdeRectToStr             ( RECT * );
-extern void     WdeStrToRect             ( char *, RECT * );
-extern void     WdeShowObjectWindow      ( HWND, Bool );
-extern void     WdeBringWindowToTop      ( HWND );
-extern Bool     WdePutObjFirst           ( OBJPTR, LIST **);
-extern Bool     WdeFindObjectsAtPt       ( POINT *, LIST **, LIST * );
-extern Bool     WdeFindSubObjects        ( SUBOBJ_REQUEST *, LIST **, LIST *);
-extern void     WdeDisableChildWindows   ( HWND );
-extern Bool     WdeProcessMouse          ( HWND, UINT, WPARAM, LPARAM );
-extern BOOL     WdeControlDefine         ( WdeDefineObjectInfo * );
-extern Bool     WdeProcessSymbolCombo    ( HWND, UINT, WPARAM, LPARAM,
-                                           WdeHashTable *, uint_16, Bool );
-extern Bool     WdeProcessHelpSymbolCombo( HWND, UINT, WPARAM, LPARAM,
-                                           WdeHashTable *, uint_32, Bool );
-extern void     WdeSetDefineObjectSymbolInfo ( WdeDefineObjectInfo *, HWND );
-extern void     WdeGetDefineObjectSymbolInfo ( WdeDefineObjectInfo *, HWND );
-extern void     WdeGetDefineObjectHelpSymbolInfo ( WdeDefineObjectInfo *, HWND );
-extern void     WdeMapWindowRect         ( HWND, HWND, RECT *);
-extern Bool     WdeIsStrSpace            ( char * );
-extern Bool     WdeWinStylesHook         ( HWND, WORD, WPARAM, LPARAM,
-                                           DialogStyle );
-extern void     WdeAddSymbolToObjectHashTable ( WdeResInfo *, char *,
-                                                WdeHashValue );
+extern UINT     WdeTranslateNCMouseMsg( UINT );
+extern void     WdeSnapPointToGrid( POINT *pt );
+extern OBJPTR   WdeGetNextObject( Bool, OBJPTR, OBJPTR );
+extern OBJPTR   WdeCloneObject( OBJPTR, POINT * );
+extern void     WdeFreeControlIDs( void );
+extern uint_16  WdeGetNextControlID( void );
+extern uint_32  WdeHammingDistance( uint_32, uint_32 );
+extern Bool     WdeIsClassDefined( char * );
+extern char     *WdeRectToStr( RECT * );
+extern void     WdeStrToRect( char *, RECT * );
+extern void     WdeShowObjectWindow( HWND, Bool );
+extern void     WdeBringWindowToTop( HWND );
+extern Bool     WdePutObjFirst( OBJPTR, LIST **);
+extern Bool     WdeFindObjectsAtPt( POINT *, LIST **, LIST * );
+extern Bool     WdeFindSubObjects( SUBOBJ_REQUEST *, LIST **, LIST * );
+extern void     WdeDisableChildWindows( HWND );
+extern Bool     WdeProcessMouse( HWND, UINT, WPARAM, LPARAM );
+extern BOOL     WdeControlDefine( WdeDefineObjectInfo * );
+extern Bool     WdeProcessSymbolCombo( HWND, UINT, WPARAM, LPARAM, WdeHashTable *, uint_16, Bool );
+extern Bool     WdeProcessHelpSymbolCombo( HWND, UINT, WPARAM, LPARAM, WdeHashTable *, uint_32, Bool );
+extern void     WdeSetDefineObjectSymbolInfo( WdeDefineObjectInfo *, HWND );
+extern void     WdeGetDefineObjectSymbolInfo( WdeDefineObjectInfo *, HWND );
+extern void     WdeGetDefineObjectHelpSymbolInfo( WdeDefineObjectInfo *, HWND );
+extern void     WdeMapWindowRect( HWND, HWND, RECT * );
+extern Bool     WdeIsStrSpace( char * );
+extern Bool     WdeWinStylesHook( HWND, UINT, WPARAM, LPARAM, DialogStyle );
+extern void     WdeAddSymbolToObjectHashTable( WdeResInfo *, char *, WdeHashValue );
 
-extern void     WdeEXSetDefineInfo ( WdeDefineObjectInfo *, HWND );
-extern void     WdeEXGetDefineInfo ( WdeDefineObjectInfo *, HWND );
+extern void     WdeEXSetDefineInfo( WdeDefineObjectInfo *, HWND );
+extern void     WdeEXGetDefineInfo( WdeDefineObjectInfo *, HWND );
 
 #endif
