+++ bld/trap/lcl/win/c/ints.c
@@ -194,7 +194,7 @@ void __export FAR PASCAL DoneWithInterrupt( LPVOID data )
         memcpy( &SaveEAX, data, sizeof( interrupt_struct ) );
     }
     IntAccessed--;
-    if( IntAccessed <= 0 ) {
+    if( (int)IntAccessed <= 0 ) {
         IDTInit( IDTSel );
         HookIDT( ReflectInt1Int3 );
         OurOwnInt = 0;
